pyproject.toml
[project]
name = "tennis-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "psycopg[binary]>=3.3.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "types-python-jose>=3.5.0.20250531",
    "types-passlib>=1.7.7.20260211",
]
